chatbot-backend/requirements.txt
# requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
python-dotenv>=0.20
google-cloud-aiplatform>=1.87.0
google-cloud-storage>=2.5
google-cloud-speech>=2.20
google-cloud-texttospeech>=2.14
google-cloud-discoveryengine>=0.1.0
Flask-Cors>=3.0
uuid
beautifulsoup4>=4.10
lxml
python-docx>=1.0
pdfminer.six>=20221105
chardet>=3.0
werkzeug
requests
Flask-Login>=0.6.2
Flask-JWT-Extended>=4.5.2
Flask-Bcrypt>=1.0.1
itsdangerous>=2.1.2
PyJWT>=2.7.0
celery>=5.3
redis>=5.0
Flask-Limiter>=3.0
Flask-Mail>=0.9
pydub
Pillow
html5lib
validators
langdetect
googletrans
google-generativeai>=0.4.0
tiktoken
google-cloud-translate
Flask-CORS
langchain-text-splitters
rank_bm25>=0.2.2
sentence-transformers
